+++ src/components/manage/Blocks/Image/View.jsx
@@ -7,9 +7,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import cx from 'classnames';
-import { settings } from '~/config';
 
-import { flattenToAppURL } from '@plone/volto/helpers';
+import { flattenToAppURL, isInternalURL } from '@plone/volto/helpers';
 
 /**
  * View image block class.
@@ -34,18 +33,16 @@ const View = ({ data, detached }) => (
             <img
               className={cx({ 'full-width': data.align === 'full' })}
               src={
-                data.url.includes(settings.apiPath)
-                  ? `${flattenToAppURL(data.url)}/@@images/image`
+                isInternalURL(data.url)
+                  ? // Backwards compat in the case that the block is storing the full server URL
+                    `${flattenToAppURL(data.url)}/@@images/image`
                   : data.url
               }
               alt={data.alt || ''}
             />
           );
           if (data.href) {
-            if (
-              (data.href.startsWith('http') || data.href.startsWith('https')) &&
-              !data.href.includes(settings.apiPath)
-            ) {
+            if (!isInternalURL(data.href)) {
               return (
                 <a
                   target={data.openLinkInNewTab ? '_blank' : null}
@@ -57,7 +54,7 @@ const View = ({ data, detached }) => (
             } else {
               return (
                 <Link
-                  to={data.href.replace(settings.apiPath, '')}
+                  to={flattenToAppURL(data.href)}
                   target={data.openLinkInNewTab ? '_blank' : null}
                 >
                   {image}