+++ CHANGELOG.md
@@ -7,8 +7,12 @@
 ### Feature
 
 - Added recurrence widget @giuliaghisini
+- Support loading additional reducers from the ``config.addonReducers`` key,
+  to allow addons to provide their own reducers @tiberiuichim
 
 ### Bugfix
+
+- Do not store the server information of the image block in the block @sneridagh
 
 ### Internal